t2

+++ training4.2_GitTest/src/Student.java
@@ -1,10 +1,10 @@
 
 public class Student {
 
-	private int age;
+	private long age;
 	private String name;
 	
-	public int getAge() {
+	public long getAge() {
 		return age;
 	}
 	public void setAge(int age) {
